+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/bot/usecase/HandleButtonInteractionUseCase.kt
@@ -73,6 +73,10 @@ internal class HandleButtonInteractionUseCase(
                 interaction.deferPublicMessageUpdate()
                 redirect(button.channelId, message)
             }
+            is DiscordButton.Text -> {
+                val response = interaction.deferPublicResponse()
+                text(interaction = interaction, response = response, text = button.text)
+            }
             else -> {
                 Result.Error(BotError.BotLogicError("Invalid button action"))
             }
@@ -205,5 +209,19 @@ internal class HandleButtonInteractionUseCase(
             Result.Error(BotError.Unknown(e.toString()))
         }
         return result
+    }
+
+    private suspend fun text(
+        interaction: ButtonInteraction,
+        response: DeferredPublicMessageInteractionResponseBehavior,
+        text: String,
+    ): EmptyResult<BotError> {
+        response.respond {
+            content = buildString {
+                appendLine(interaction.user.mention)
+                append(text)
+            }
+        }
+        return Result.Success(Unit)
     }
 }