+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/wikiDragDown/DragDownWikiDiscordFeature.kt
@@ -25,6 +25,8 @@ import io.github.sophon.discord.feat.core.usecase.GetCharactersUseCase
 import io.github.sophon.discord.feat.core.usecase.GetMoveUseCase
 import io.github.sophon.discord.feat.core.usecase.GetMovesUseCase
 import io.github.sophon.discord.feat.core.usecase.SyncWikiDataUseCase
+import io.github.sophon.discord.feat.core.util.aggregateCharacters
+import io.github.sophon.discord.feat.core.util.firstMatchingWikiMoves
 import io.github.sophon.discord.util.toButtons
 import io.github.sophon.discord.util.withWiki
 import io.github.sophon.integration.model.Source
@@ -133,10 +135,17 @@ internal class DragDownWikiDiscordFeature(
         return result
     }
 
+    override suspend fun getAllCharacters() =
+        aggregateCharacters(wikiClientMap, getCharactersUseCase)
+
     override suspend fun getMoveList(
         command: Command,
         characterId: String,
     ): Result<List<Move>, BotError> {
+        if (command == Command.Fd) {
+            val moves = firstMatchingWikiMoves(wikiClientMap, getMovesUseCase, characterId)
+            return moves
+        }
         val game = when (command) {
             Command.FdROA -> Game.ROA2
             else -> return Result.Error(BotError.BotLogicError(command.name, ""))