+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/wikiWavu/WavuEmbeds.kt
@@ -35,10 +35,6 @@ internal fun wavuMoveEmbed(
 
     createNotes(move)
 
-    move.urls.videoId?.let { url ->
-        optionalField(name = "Video", value = "[Link](${url})", inline = false)
-    }
-
     featureFooter(featureInfo)
 }