+++ feat/wikiDragDown/src/commonTest/kotlin/io/github/sophon/wikidragdown/data/MoveMapperTest.kt
@@ -12,7 +12,7 @@ internal class MoveMapperTest {
     fun `mapping handles basic move`() {
         // given
         val dto = MoveSource.olympiaBair
-        val character = CharacterSource.olympia
+        val character = DomainCharacterSource.olympia
         val expected = Move(
             id = "olympia_bair",
             characterId = character.id,
@@ -22,7 +22,7 @@ internal class MoveMapperTest {
             name = dto.attack,
             input = dto.attack.orEmpty().lowercase(),
             urls = Move.Urls(
-                wikiUrl = "${WIKI_BASE_URL}/${character.remoteQueryId}",
+                wikiUrl = "https://dragdown.wiki/wiki/RoA2/Olympia",
             ),
             roa2Properties = Move.Roa2Properties(
                 caption = dto.caption,
@@ -52,7 +52,7 @@ internal class MoveMapperTest {
         )
 
         // when
-        val result = dto.toDomain(CharacterSource.olympia, mapOf())
+        val result = dto.toDomain(DomainCharacterSource.olympia, mapOf())
 
         //then
         assertThat(result).isEqualTo(expected)
@@ -62,7 +62,7 @@ internal class MoveMapperTest {
     fun `mapping handles parentheses special`() {
         // given
         val dto = MoveSource.forsburnDSpecialEmptyInhale
-        val character = CharacterSource.forsburn
+        val character = DomainCharacterSource.forsburn
         val expected = Move(
             id = "forsburn_dspecialemptyinhale",
             characterId = character.id,
@@ -72,7 +72,7 @@ internal class MoveMapperTest {
             name = dto.attack,
             input = "dspecialemptyinhale",
             urls = Move.Urls(
-                wikiUrl = "${WIKI_BASE_URL}/${character.remoteQueryId}",
+                wikiUrl = "https://dragdown.wiki/wiki/RoA2/Forsburn",
             ),
             roa2Properties = Move.Roa2Properties(
                 mode = "emptyinhale",
@@ -103,7 +103,7 @@ internal class MoveMapperTest {
         )
 
         // when
-        val result = dto.toDomain(CharacterSource.forsburn, mapOf())
+        val result = dto.toDomain(DomainCharacterSource.forsburn, mapOf())
 
         //then
         assertThat(result).isEqualTo(expected)
@@ -113,7 +113,7 @@ internal class MoveMapperTest {
     fun `mapping handles html formatting`() {
         // given
         val dto = MoveSource.rannoUspecialDivekick
-        val character = CharacterSource.ranno
+        val character = DomainCharacterSource.ranno
         val expected = Move(
             id = "ranno_uspecialdivekick",
             characterId = character.id,
@@ -124,7 +124,7 @@ internal class MoveMapperTest {
             input = "uspecialdivekick",
             cancel = "Double Jump & Wall Jump: 43+\nLedge Grab: 57+",
             urls = Move.Urls(
-                wikiUrl = "${WIKI_BASE_URL}/${character.remoteQueryId}",
+                wikiUrl = "https://dragdown.wiki/wiki/RoA2/Ranno",
             ),
             roa2Properties = Move.Roa2Properties(
                 mode = "divekick",
@@ -155,7 +155,7 @@ internal class MoveMapperTest {
         )
 
         // when
-        val result = dto.toDomain(CharacterSource.ranno, mapOf())
+        val result = dto.toDomain(DomainCharacterSource.ranno, mapOf())
 
         //then
         assertThat(result).isEqualTo(expected)
@@ -165,7 +165,7 @@ internal class MoveMapperTest {
     fun `mapping handles FADC special`() {
         // given
         val dto = MoveSource.olympiaDSpecialFADC
-        val character = CharacterSource.olympia
+        val character = DomainCharacterSource.olympia
         val expected = Move(
             id = "olympia_dspecialfadc",
             characterId = character.id,
@@ -176,7 +176,7 @@ internal class MoveMapperTest {
             input = "dspecialfadc",
             cancel = "Ledge Grab: 6",
             urls = Move.Urls(
-                wikiUrl = "${WIKI_BASE_URL}/${character.remoteQueryId}",
+                wikiUrl = "https://dragdown.wiki/wiki/RoA2/Olympia",
             ),
             notes = listOf(
                 "This move has a 36 frame [cooldown](https://dragdown.wiki/wiki/RoA2/System_Mechanics/Misc#Cooldowns) once endlag begins.",
@@ -209,7 +209,7 @@ internal class MoveMapperTest {
         )
 
         // when
-        val result = dto.toDomain(CharacterSource.olympia, mapOf())
+        val result = dto.toDomain(DomainCharacterSource.olympia, mapOf())
 
         //then
         assertThat(result).isEqualTo(expected)
@@ -392,7 +392,7 @@ private object MoveSource {
     )
 }
 
-private object CharacterSource {
+private object DomainCharacterSource {
     val olympia = Character(
         id = "olympia",
         displayName = "Olympia",