+++ feat/wikiDreamCancel/src/commonMain/kotlin/io/github/sophon/dreamcancel/data/CharacterMapper.kt
@@ -1,5 +1,6 @@
 package io.github.sophon.dreamcancel.data
 
+import io.github.sophon.core.featureConfig.model.Game
 import io.github.sophon.core.util.cleanHtml
 import io.github.sophon.core.util.createAliases
 import io.github.sophon.core.util.removeAccents
@@ -8,6 +9,7 @@ import io.github.sophon.dreamcancel.domain.FEATURE_URL
 
 internal fun String.toDomain(
     gameId: String,
+    imageUrlMap: Map<String, String>,
 ): Character {
     val idName = this
         .cleanHtml()
@@ -17,6 +19,8 @@ internal fun String.toDomain(
         .lowercase()
     val displayName = this.cleanHtml()
     val queryName = this.createQueryName()
+    val iconKeys = listOf(this.substringBefore(" "), this.substringAfterLast(" "))
+    val iconUrl = iconKeys.firstNotNullOfOrNull { imageUrlMap[it] } ?: Game.fromId(gameId)?.iconUrl
 
     val char = Character(
         id = idName,
@@ -24,6 +28,9 @@ internal fun String.toDomain(
         aliasList = displayName.createAliases(),
         remoteQueryId = queryName,
         wikiUrl = "$FEATURE_URL/$gameId/$queryName",
+        images = Character.Images(
+            iconUrl = iconUrl,
+        )
     )
 
     return char
@@ -35,23 +42,4 @@ internal fun String.createQueryName(): String {
         .removeAccents()
         .split(' ')
         .joinToString("_")
-}
-
-internal fun String.createThumbnailUrl(gameId: String): String? {
-    val prefix: String
-    val suffix: String
-
-    when (gameId) {
-        DreamCancelTables.TABLE_COTW_MOVES -> {
-            prefix = "FF_COTW"
-            suffix = "Icon.png"
-        }
-        DreamCancelTables.TABLE_KOF15_MOVES -> {
-            prefix = "KOFXV"
-            suffix = "Portrait.png"
-        }
-        else -> return null
-    }
-
-    return "${prefix}_${this}_${suffix}"
 }