+++ feat/wikiDustLoop/src/commonMain/kotlin/io/github/sophon/wikidustloop/integration/DustLoopFeatureInfo.kt
@@ -3,6 +3,7 @@ package io.github.sophon.wikidustloop.integration
 import io.github.sophon.core.featureConfig.model.FeatureInfo
 import io.github.sophon.core.featureConfig.model.Game
 import io.github.sophon.dustloop.BuildKonfig
+import io.github.sophon.wikidustloop.domain.FEATURE_FEEDBACK_DISCORD_ID
 import io.github.sophon.wikidustloop.domain.FEATURE_IMG_URL
 import io.github.sophon.wikidustloop.domain.FEATURE_NAME
 import io.github.sophon.wikidustloop.domain.FEATURE_URL
@@ -17,7 +18,9 @@ object DustLoopFeatureInfo {
             Game.DBFZ,
             Game.GBVSR,
             Game.BBCF,
+            Game.MTFS,
         ),
         version = BuildKonfig.VERSION,
+        feedbackDiscordChannelId = FEATURE_FEEDBACK_DISCORD_ID,
     )
 }